:recycle: (spikes): boost::ut - refactor tests in separated .cpp files
- separate tests in different files
- create small main.cpp that can be used anywhere
- update CMakeLists.txt file
- add config.h
- use MACROS for start/end statistics
