+++ spikes/lk_boost_ut/utils.h
@@ -8,38 +8,34 @@
 
 #include "LogKit.h"
 
-namespace utils {
+namespace utils::time {
 
 using namespace leka;
 using namespace std::chrono;
 
-namespace time {
+inline auto start = rtos::Kernel::Clock::now();
+inline auto stop  = rtos::Kernel::Clock::now();
+inline auto delta = static_cast<int>((stop - start).count());
 
-	auto start = rtos::Kernel::Clock::now();
-	auto stop  = rtos::Kernel::Clock::now();
-	auto delta = static_cast<int>((stop - start).count());
+// namespace time
 
-}	// namespace time
+}	// namespace utils::time
 
-inline void start(const char *msg)
-{
-	logger::init();
+// NOLINTNEXTLINE
+#define utils_start(msg)                                                                                               \
+	do {                                                                                                               \
+		logger::init();                                                                                                \
+		log_ll("\n", 1);                                                                                               \
+		log_info("Starting %s...\n", msg);                                                                             \
+		rtos::ThisThread::sleep_for(1s);                                                                               \
+		utils::time::start = rtos::Kernel::Clock::now();                                                               \
+	} while (0)
 
-	log_ll("\n", 1);
-	log_info("Starting %s...\n", msg);
-
-	rtos::ThisThread::sleep_for(1s);
-
-	time::start = rtos::Kernel::Clock::now();
-}
-
-inline void end()
-{
-	time::stop	= rtos::Kernel::Clock::now();
-	time::delta = static_cast<int>((time::stop - time::start).count());
-
-	log_ll("\n", 1);
-	log_info("End of tests (%i ms total)", time::delta);
-}
-
-}	// namespace utils
+// NOLINTNEXTLINE
+#define utils_end()                                                                                                    \
+	do {                                                                                                               \
+		utils::time::stop  = rtos::Kernel::Clock::now();                                                               \
+		utils::time::delta = static_cast<int>((utils::time::stop - utils::time::start).count());                       \
+		log_ll("\n", 1);                                                                                               \
+		log_info("End of tests (%i ms total)", utils::time::delta);                                                    \
+	} while (0)
